+++ torchci/components/TorchAgentPage.tsx
@@ -1,17 +1,9 @@
-import ThumbDownIcon from "@mui/icons-material/ThumbDown";
-import ThumbUpIcon from "@mui/icons-material/ThumbUp";
-import {
-  Box,
-  Button,
-  IconButton,
-  Tooltip,
-  Typography,
-  useTheme,
-} from "@mui/material";
+import { Box, Button, Typography, useTheme } from "@mui/material";
 import { useSession } from "next-auth/react";
 import { useEffect, useRef, useState } from "react";
 import AISpinner from "./AISpinner";
 import { ChatHistorySidebar } from "./TorchAgentPage/ChatHistorySidebar";
+import { FeedbackButtons } from "./TorchAgentPage/FeedbackButtons";
 import { GrafanaEmbed } from "./TorchAgentPage/GrafanaEmbed";
 import { HeaderSection } from "./TorchAgentPage/HeaderSection";
 import {
@@ -83,10 +75,6 @@ export const TorchAgentPage = () => {
   const [debugVisible, setDebugVisible] = useState(false);
   const [currentSessionId, setCurrentSessionId] = useState<string | null>(null);
   const [feedbackVisible, setFeedbackVisible] = useState(false);
-  const [feedbackSubmitted, setFeedbackSubmitted] = useState<number | null>(
-    null
-  );
-  const [feedbackSelected, setFeedbackSelected] = useState<number | null>(null);
 
   // Chat history state
   const [chatHistory, setChatHistory] = useState<ChatSession[]>([]);
@@ -102,11 +90,6 @@ export const TorchAgentPage = () => {
   const calculateTotalTokens = useTokenCalculator();
   const { showScrollButton, scrollToBottomAndEnable, resetAutoScroll } =
     useAutoScroll(isLoading, parsedResponses);
-
-  // Debug: Track feedbackVisible changes
-  useEffect(() => {
-    console.log("feedbackVisible changed to:", feedbackVisible);
-  }, [feedbackVisible]);
 
   const handleQueryChange = (event: React.ChangeEvent<HTMLInputElement>) => {
     setQuery(event.target.value);
@@ -139,8 +122,6 @@ export const TorchAgentPage = () => {
     setParsedResponses([]);
     setResponse("");
     setError("");
-    setFeedbackSubmitted(null);
-    setFeedbackSelected(null);
 
     try {
       const response = await fetch(
@@ -290,12 +271,6 @@ export const TorchAgentPage = () => {
 
         setSelectedSession(sessionId);
         setCurrentSessionId(sessionId);
-        console.log("sessiondata:", sessionData);
-        console.log("status:", sessionData.status);
-        console.log(
-          "setting feedbackVisible to:",
-          sessionData.status === "completed"
-        );
         setFeedbackVisible(sessionData.status === "completed");
         // Don't override the response we set above for the debug view
       } else {
@@ -318,8 +293,6 @@ export const TorchAgentPage = () => {
     setSelectedSession(null);
     setCurrentSessionId(null);
     setFeedbackVisible(false);
-    setFeedbackSubmitted(null);
-    setFeedbackSelected(null);
     setError("");
     setTotalTokens(0);
     setCompletedTokens(0);
@@ -831,25 +804,6 @@ export const TorchAgentPage = () => {
     }
   };
 
-  const sendFeedback = async (value: number) => {
-    if (!currentSessionId) return;
-    try {
-      await fetch("/api/torchagent-feedback", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ sessionId: currentSessionId, feedback: value }),
-      });
-      setFeedbackSelected(value);
-      setFeedbackSubmitted(value);
-      // Hide confirmation message after 5 seconds, but keep button selected
-      setTimeout(() => {
-        setFeedbackSubmitted(null);
-      }, 3000);
-    } catch (e) {
-      console.error("Failed to send feedback", e);
-    }
-  };
-
   // Authentication check
   if (session.status === "loading") {
     return (
@@ -1057,95 +1011,10 @@ export const TorchAgentPage = () => {
                   {formatTokenCount(completedTokens)} tokens
                 </Typography>
               )}
-              {feedbackVisible && (
-                <Box sx={{ ml: 2, position: "relative" }}>
-                  <Tooltip title="This session was helpful and gave me what I asked for">
-                    <IconButton
-                      color="primary"
-                      size="small"
-                      onClick={() => sendFeedback(1)}
-                      sx={{
-                        backgroundColor:
-                          feedbackSelected === 1
-                            ? "primary.main"
-                            : "transparent",
-                        color:
-                          feedbackSelected === 1
-                            ? "primary.contrastText"
-                            : "primary.main",
-                        "&:hover": {
-                          backgroundColor:
-                            feedbackSelected === 1
-                              ? "primary.dark"
-                              : "primary.light",
-                        },
-                      }}
-                    >
-                      <ThumbUpIcon fontSize="small" />
-                    </IconButton>
-                  </Tooltip>
-                  <Tooltip title="This output is not what I asked">
-                    <IconButton
-                      color="primary"
-                      size="small"
-                      onClick={() => sendFeedback(-1)}
-                      sx={{
-                        backgroundColor:
-                          feedbackSelected === -1
-                            ? "primary.main"
-                            : "transparent",
-                        color:
-                          feedbackSelected === -1
-                            ? "primary.contrastText"
-                            : "primary.main",
-                        "&:hover": {
-                          backgroundColor:
-                            feedbackSelected === -1
-                              ? "primary.dark"
-                              : "primary.light",
-                        },
-                      }}
-                    >
-                      <ThumbDownIcon fontSize="small" />
-                    </IconButton>
-                  </Tooltip>
-                  {feedbackSubmitted !== null && (
-                    <Box
-                      sx={{
-                        position: "absolute",
-                        top: -40,
-                        left: "50%",
-                        transform: "translateX(-50%)",
-                        backgroundColor: "success.main",
-                        color: "success.contrastText",
-                        px: 2,
-                        py: 1,
-                        borderRadius: 1,
-                        fontSize: "0.875rem",
-                        fontWeight: "medium",
-                        boxShadow: 2,
-                        zIndex: 1000,
-                        whiteSpace: "nowrap",
-                        "&::after": {
-                          content: '""',
-                          position: "absolute",
-                          top: "100%",
-                          left: "50%",
-                          transform: "translateX(-50%)",
-                          width: 0,
-                          height: 0,
-                          borderLeft: "6px solid transparent",
-                          borderRight: "6px solid transparent",
-                          borderTop: "6px solid",
-                          borderTopColor: "success.main",
-                        },
-                      }}
-                    >
-                      Feedback recorded, thank you!
-                    </Box>
-                  )}
-                </Box>
-              )}
+              <FeedbackButtons
+                sessionId={currentSessionId}
+                visible={feedbackVisible}
+              />
             </Box>
           )
         )}
